flask/requirements.txt
flask==2.3.3
flask-cors==3.0.10
psycopg2-binary